Use long int explicitly for event index as we can analyse very large numbers of events.
Former-commit-id: 5106892768ad6edcd1b6f2f4adf64849ab820dec

+++ analysis/Analysis.cc
@@ -109,7 +109,7 @@ void Analysis::PreparePerEntryHistograms()
     }
 }
 
-void Analysis::AccumulatePerEntryHistograms(const int& entryNumber)
+void Analysis::AccumulatePerEntryHistograms(const long int& entryNumber)
 {
   auto c = Config::Instance();
   if (c)
