+++ analysis/PerEntryHistogram.cc
@@ -87,7 +87,7 @@ PerEntryHistogram::~PerEntryHistogram()
   delete accumulator;
 }
       
-void PerEntryHistogram::AccumulateCurrentEntry(const int& entryNumber)
+void PerEntryHistogram::AccumulateCurrentEntry(const long int& entryNumber)
 {  
   // Fill the temporary histogram with 1 event - the current one
   // This is used as it doesn't matter if the variable is a vector
